pyproject.toml
[tool.poetry]
name = "anime-mvp"
version = "0.1.0"
description = "Anime data pipeline with Jikan API ingestion, AWS processing, and ML recommendations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
requests = "^2.31.0"
boto3 = "^1.34.0"
pandas = "^2.1.4"
pyspark = "^3.5.0"
awswrangler = "^3.5.0"
python-dotenv = "^1.0.0"
s3fs = "^2023.12.0"
pyarrow = "^14.0.0"
pyautogen = "^0.10.0"
openai = "^1.109.1"
streamlit = "^1.50.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
moto = "^4.2.12"
pytest-mock = "^3.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
fetch-anime = "src.ingestion.fetch_jikan:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"